Fixed Compound Index Strategy
getQueryRanges(final MultiDimensionalNumericData indexedRange
,final int maxEstimatedRangeDecomposition) now handles negative
values for maxEstimatedRangeDecomposition by performing a full
decomposition.

Both getMaxQueryRanges(...) and getInsertionIds(...) take the
actual number of results from the call to the first sub-strategy
into account when calling the second strategy. For example, if the
first sub-strategy returned a single range, the second sub-strategy
can use the full maxRangeDecomposition (since the product of the two
will be maxRangeDecomposition).

getQueryRanges(...) uses the ceiling for computing the maximum estimated
range decomposition per strategy since the parameter is not strictly
enforced. getInsertionIds(...) uses the floor for each strategy
to reduce replication.