Implemented a Simple NumericIndexStrategy
Created a simple 1-dimensional NumericIndexStrategy that
can index integer values. The strategy doesn't use any
binning. The ids are simply the byte array values of
the numbers. This index strategy won't work well for
inserting ranges since the data will be replicated for
each integer value in the range.

The SimpleNumericIndexStrategy is an abstract class that
uses a Short, Integer, or Long lexicoder to encode/decode
numbers. This allows for adjusting the size of the keys
(i.e. a 16 or 32 bit integer might give a large enough
range for some applications) to save space in the row ID.
Extended by SimpleShortIndexStrategy (for 16 bit keys),
SimpleIntegerIndexStrategy (for 32 bit keys), and
SimpleLongIndexStrategy (for 64 bit keys).

getQueryRanges always returns 1 range since the sort order of the
bytes is the same as the sort order of the values. This means that
a contiguous range of values can be represented by a single
contiguous ByteArrayRange.

getInsertionIds returns a single insertionId if you are inserting
a point. If you insert a range, you'll get back a list with the
size of the number of integers in the range (because there is no
binning). That is not the intended use of this index strategy.