Changed the featuretype for TRACK to LineString to allow the STANAG 4676 data to be scanned.
While scanning 4676 line featuretype data in Accumulo, the iterator throws an exception due to a bad CQL filter string.

This is where the error occurs:

https://github.com/geoserver/geoserver/blob/master/src/wms/src/main/java/org/geoserver/wms/featureinfo/FeatureInfoStylePreprocessor.java#L218

It is because the geometry is of binding class Geometry.class (rather than LineString.class) so being specific and using LineString.class on our ingest should be a reasonable workaround.