Issue #659, #658, #660.  Allow the index writer to to multiple indices.  Retain the index/adapter assocation to identify possibile indices to query for a given adapter and to enfore a consistent set of indices to write for a given adapter across multiple ingest runs or WFS-T transactions.  Lastly, eliminate the redudant visibility constraint when creating an index writer
Issue #659, #658, #660.  Allow the index writer to to multiple indices.  Retain the index/adapter assocation to identify possibile indices to query for a given adapter and to enfore a consistent set of indices to write for a given adapter across multiple ingest runs or WFS-T transactions.  Lastly, eliminate the redudant visibility constraint when creating an index writer.

GEOWAVE-669: moved geowavedemoapp to exmaples so that it could use adapter dependencies; demo app works in windows now

Issue #659, #658, #660.  Allow the index writer to to multiple indices.  Retain the index/adapter assocation to identify possibile indices to query for a given adapter and to enfore a consistent set of indices to write for a given adapter across multiple ingest runs or WFS-T transactions.  Lastly, eliminate the redudant visibility constraint when creating an index writer

Issue #662.  Partition based on boundary of polygon

GEOWAVE-669: moved geowavedemoapp to exmaples so that it could use adapter dependencies; demo app works in windows now

GEOWAVE-673: added index as optional parameter to KDE

merged