Enable batch correlation of uploaded files
This change allows an external consumer of the REST API to submit files in a
batch and ingest them in a single shot as opposed to submitting a single
ingest job per file.

Currently `geowave ingest localToGw` only accepts a single file or a directory
as the source parameter.  Doing "batch" ingests by transferring files via
`/restservices/v0/fileupload` would require invoking `localToGw` once for
every file in the batch which seems like it would require extra bookkeeping to
track and merge the operation statuses.

Refer to G3-297