Initialize watchdog only when needed, fix kudu options factory
If the watchdog is initialized too early, it will hang if we attempt to
check the status of it, since it will call wait() to allow a process to
start. In the case we check for a process running, the watchdog will
hang because it did not start in the first place.

Signed-off-by: Danny Qiu <dqiu55@gmail.com>