Replace old statistics system
The previous statistics system was difficult to extend and utilize effectively. This pull request replaced the system with a new system that decouples statistics from data type adapter implementations and allows third party developers to easily add their own statistics and binning strategies without having to write new adapters. It includes refreshed statistics APIs and CLI commands that take advantage of all the new capabilities.

Note: Due of the scale of this change, existing statistics are no longer compatible and will need to be recalculated.
Signed-off-by: Johnathan Garrett <jd@prominentedge.com>