Replace old statistics system (#1792)
* Replace old statistics system

The previous statistics system was difficult to extend and utilize effectively. This pull request replaced the system with a new system that decouples statistics from data type adapter implementations and allows third party developers to easily add their own statistics and binning strategies without having to write new adapters. It includes refreshed statistics APIs and CLI commands that take advantage of all the new capabilities.

Note: Due of the scale of this change, existing statistics are no longer compatible and will need to be recalculated.
Signed-off-by: Johnathan Garrett <jd@prominentedge.com>

* Add more statistic types and beef up testing

Signed-off-by: Rich Fecher <richard.fecher@maxar.com>

* Fix issue with multiple threads updating a statistic simultaneously

Signed-off-by: Johnathan Garrett <jd@prominentedge.com>

* better logic for multifields and tests

Signed-off-by: Rich Fecher <richard.fecher@maxar.com>

* Update documentation with latest binning strategy changes

Signed-off-by: Johnathan Garrett <jd@prominentedge.com>

* Fix formatting errors

Signed-off-by: Johnathan Garrett <jd@prominentedge.com>

* FIx issue with merging stats on Kudu

Signed-off-by: Johnathan Garrett <jd@prominentedge.com>

Co-authored-by: Rich Fecher <richard.fecher@maxar.com>